online_inference/requirements.txt
pandas==1.2.4
pytest==6.2.3
uvicorn[standard]==0.13.4
fastapi==0.64.0
python-dotenv==0.17.1
scikit-learn==0.24.1
omegaconf==2.0.6
git+git://github.com/polikutinevgeny/cloudpickle.git@206-deep-serialization
